restructure project to allow more artifacts to be built (e.g. a CLI)
This commit merely moves files around to form this structure:

```
client/              # A ChaosSearch client library
terraformprovider/   # A Terraform provider (uses client)
cli/                 # A CLI (uses client)
```
