+++ .github/workflows/release.yml
@@ -11,7 +11,7 @@
 # You will need to pass the `--batch` flag to `gpg` in your signing step 
 # in `goreleaser` to indicate this is being used in a non-interactive mode.
 #
-name: release
+name: release Terraform provider
 on:
   push:
     tags:
@@ -44,6 +44,7 @@ jobs:
         with:
           version: latest
           args: release --rm-dist
+          workdir: terraformprovider
         env:
           GPG_FINGERPRINT: ${{ steps.import_gpg.outputs.fingerprint }}
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
