+++ client/configuration.go
