+++ client/operation_delete_object_group.go
