+++ client/operation_read_object_group.go
