+++ client/operation_update_indexing_state.go
