+++ client/operation_update_object_group.go
