+++ go.mod
@@ -1,11 +1,11 @@
-module terraform-provider-chaossearch
+module chaossearch
 
 go 1.14
 
 require (
-	github.com/antihax/optional v1.0.0
 	github.com/aws/aws-sdk-go v1.25.3
 	github.com/hashicorp/terraform-plugin-sdk/v2 v2.2.0
 	golang.org/x/net v0.0.0-20201110031124-69a78807bb2b // indirect
-	golang.org/x/oauth2 v0.0.0-20201109201403-9fd604954f58
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
+	google.golang.org/genproto v0.0.0-20200825200019-8632dd797987 // indirect
 )
