+++ terraformprovider/main.go
@@ -3,14 +3,12 @@ package main
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
-
-	"terraform-provider-chaossearch/chaossearch"
 )
 
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
 		ProviderFunc: func() *schema.Provider {
-			return chaossearch.Provider()
+			return Provider()
 		},
 	})
 }
