+++ terraformprovider/resource_indexing_state.go
@@ -1,8 +1,8 @@
-package chaossearch
+package main
 
 import (
+	"chaossearch/client"
 	"context"
-	"terraform-provider-chaossearch/chaossearch/client"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
